ScatterNNlib

ScatterNNlib.gather

ScatterNNlib.save_div

ScatterNNlib.scatter_add!

ScatterNNlib.scatter_div!

ScatterNNlib.scatter_max!

ScatterNNlib.scatter_max!

ScatterNNlib.scatter_mean!

ScatterNNlib.scatter_mean!

ScatterNNlib.scatter_min!

ScatterNNlib.scatter_min!

ScatterNNlib.scatter_mul!

ScatterNNlib.scatter_sub!

ScatterNNlib.gather — Method

Inverse operation of scatter

ScatterNNlib.save_div — Method

save_div(x, y)

Savely divde x by y. If y is zero, return x directly.

ScatterNNlib.scatter_add! — Function

scatter_add!(ys, us, xs)

Scatter addition operation.

ys[:, xs[k]...] = ys[:, xs[k]...] .+ us[:, k...]

Arguments

ys: the destination for us to aggregate to. This argument will be mutated.

us: the source data for aggregating.

xs: the mapping for aggregation from source (index) to destination (value).

The index of xs is corresponding to the index of us and the value of xs is corresponding to the index of ys. The value of xs can be Int or Tuple type.

The dimension of us must equal to one plus dimension of xs. ys, us and xs must be supported array type and be the same type.Array, StaticArray and CuArray are currently supported.

ScatterNNlib.scatter_div! — Function

scatter_div!(ys, us, xs)

Scatter division operation.

ys[:, xs[k]...] = ys[:, xs[k]...] ./ us[:, k...]

Arguments

ys: the destination for us to aggregate to. This argument will be mutated.

us: the source data for aggregating.

xs: the mapping for aggregation from source (index) to destination (value).

The index of xs is corresponding to the index of us and the value of xs is corresponding to the index of ys. The value of xs can be Int or Tuple type.

The dimension of us must equal to one plus dimension of xs. ys, us and xs must be supported array type and be the same type.Array, StaticArray and CuArray are currently supported.

ScatterNNlib.scatter_max! — Method

scatter_max!(ys, us, xs)

Scatter maximum operation.

ys[:, xs[k]...] = max.(ys[:, xs[k]...], us[:, k...])

Arguments

ys: the destination for us to aggregate to. This argument will be mutated.

us: the source data for aggregating.

xs: the mapping for aggregation from source (index) to destination (value).

The index of xs is corresponding to the index of us and the value of xs is corresponding to the index of ys. The value of xs can be Int or Tuple type.

The dimension of us must equal to one plus dimension of xs. ys, us and xs must be supported array type and be the same type.Array, StaticArray and CuArray are currently supported.

ScatterNNlib.scatter_mean! — Method

scatter_mean!(ys, us, xs)

Scatter mean operation.

ys[:, xs[k]...] = mean.(ys[:, xs[k]...], us[:, k...])

Arguments

ys: the destination for us to aggregate to. This argument will be mutated.

us: the source data for aggregating.

xs: the mapping for aggregation from source (index) to destination (value).

The index of xs is corresponding to the index of us and the value of xs is corresponding to the index of ys. The value of xs can be Int or Tuple type.

The dimension of us must equal to one plus dimension of xs. ys, us and xs must be supported array type and be the same type.Array, StaticArray and CuArray are currently supported.

ScatterNNlib.scatter_min! — Method

scatter_min!(ys, us, xs)

Scatter minimum operation.

ys[:, xs[k]...] = min.(ys[:, xs[k]...], us[:, k...])

Arguments

ys: the destination for us to aggregate to. This argument will be mutated.

us: the source data for aggregating.

xs: the mapping for aggregation from source (index) to destination (value).

The index of xs is corresponding to the index of us and the value of xs is corresponding to the index of ys. The value of xs can be Int or Tuple type.

The dimension of us must equal to one plus dimension of xs. ys, us and xs must be supported array type and be the same type.Array, StaticArray and CuArray are currently supported.

ScatterNNlib.scatter_mul! — Function

scatter_mul!(ys, us, xs)

Scatter multiplication operation.

ys[:, xs[k]...] = ys[:, xs[k]...] .* us[:, k...]

Arguments

ys: the destination for us to aggregate to. This argument will be mutated.

us: the source data for aggregating.

xs: the mapping for aggregation from source (index) to destination (value).

The index of xs is corresponding to the index of us and the value of xs is corresponding to the index of ys. The value of xs can be Int or Tuple type.

The dimension of us must equal to one plus dimension of xs. ys, us and xs must be supported array type and be the same type.Array, StaticArray and CuArray are currently supported.

ScatterNNlib.scatter_sub! — Function

scatter_sub!(ys, us, xs)

Scatter subtraction operation.

ys[:, xs[k]...] = ys[:, xs[k]...] .- us[:, k...]

Arguments

ys: the destination for us to aggregate to. This argument will be mutated.

us: the source data for aggregating.

xs: the mapping for aggregation from source (index) to destination (value).

The index of xs is corresponding to the index of us and the value of xs is corresponding to the index of ys. The value of xs can be Int or Tuple type.

The dimension of us must equal to one plus dimension of xs. ys, us and xs must be supported array type and be the same type.Array, StaticArray and CuArray are currently supported.