Scatter operations

Missing docstring.

Missing docstring for scatter_add!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for scatter_sub!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for scatter_mul!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for scatter_div!. Check Documenter's build log for details.

ScatterNNlib.scatter_max! — Function

scatter_max!(ys, us, xs)

Scatter maximum operation.

ys[:, xs[k]...] = max.(ys[:, xs[k]...], us[:, k...])

Arguments

ys: the destination for us to aggregate to. This argument will be mutated.

us: the source data for aggregating.

xs: the mapping for aggregation from source (index) to destination (value).

The index of xs is corresponding to the index of us and the value of xs is corresponding to the index of ys. The value of xs can be Int or Tuple type.

The dimension of us must equal to one plus dimension of xs. ys, us and xs must be supported array type and be the same type.Array, StaticArray and CuArray are currently supported.

ScatterNNlib.scatter_min! — Function

scatter_min!(ys, us, xs)

Scatter minimum operation.

ys[:, xs[k]...] = min.(ys[:, xs[k]...], us[:, k...])

Arguments

ys: the destination for us to aggregate to. This argument will be mutated.

us: the source data for aggregating.

xs: the mapping for aggregation from source (index) to destination (value).

The index of xs is corresponding to the index of us and the value of xs is corresponding to the index of ys. The value of xs can be Int or Tuple type.

The dimension of us must equal to one plus dimension of xs. ys, us and xs must be supported array type and be the same type.Array, StaticArray and CuArray are currently supported.

ScatterNNlib.scatter_mean! — Function

scatter_mean!(ys, us, xs)

Scatter mean operation.

ys[:, xs[k]...] = mean.(ys[:, xs[k]...], us[:, k...])

Arguments

ys: the destination for us to aggregate to. This argument will be mutated.

us: the source data for aggregating.

xs: the mapping for aggregation from source (index) to destination (value).

The index of xs is corresponding to the index of us and the value of xs is corresponding to the index of ys. The value of xs can be Int or Tuple type.

The dimension of us must equal to one plus dimension of xs. ys, us and xs must be supported array type and be the same type.Array, StaticArray and CuArray are currently supported.